feat: add tsc transpiling & overall clean up
