+++ package.json
@@ -5,17 +5,20 @@
     "main": "index.js",
     "type": "module",
     "scripts": {
-        "start": "node ./index.ts",
-        "start:dev": "nodemon ./index.ts"
+        "start": "nodemon ./src/index.ts",
+        "start:dev": "nodemon ./src/index.ts",
+        "start:prod": "tsc && node ./dist/index.js"
     },
-    "author": "Andrei Auchynnikau",
+    "author": "Pavel Grinkevich",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
         "nodemon": "^2.0.16",
         "ts-loader": "^9.3.0",
         "ts-node": "^10.8.1",
-        "typescript": "^4.7.4"
+        "typescript": "^4.7.4",
+        "webpack": "^5.73.0",
+        "webpack-cli": "^4.10.0"
     },
     "dependencies": {
         "@types/jimp": "^0.2.28",
